+++ src/Resources/ResourceManager/SdkClient/DeploymentScriptsSdkClient.cs
@@ -71,7 +71,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
             return psDeploymentScriptObject;
         }
 
-        public virtual PsDeploymentScriptLog GetDeploymentScriptLog(string scriptName, string resourceGroupName, int? tail)
+        public virtual PsDeploymentScriptLog GetDeploymentScriptLog(string scriptName, string resourceGroupName, int tail)
         {
             var deploymentScriptLog =
                 DeploymentScriptsClient.DeploymentScripts.GetLogsDefault(resourceGroupName, scriptName, tail);