+++ src/Resources/Resources/ChangeLog.md
@@ -21,6 +21,7 @@
 * Add message warning about view delay when creating a new Role Definition
 * Change policy cmdlets to output strongly-typed objects
 * Fixed `Remove-AzResourceGroup -Id ResourceId`[#9882]
+* Added Tail parameter to Get-AzDeploymentScriptLog and Save-AzDeploymentScriptLog cmdlets
 
 ## Version 1.13.0
 * Fixed `Get-AzResource -ResourceGroupName -Name -ExpandProperties -ResourceType` to use actual apiVersion of resources instead of default apiVersion [#11267]